+++ docs/changelog.rst
@@ -34,6 +34,10 @@ Bug fixes:
   as a numpy array. Update ``librosa`` dependency constraint to prevent similar
   issues in the future.
   :bug:`5289`
+* :doc:`plugins/lyrics`: Fix fetching lyrics from ``lrclib`` source. Instead of
+  attempting to fetch lyrics for a specific album, artist, title and duration
+  combination, the plugin now performs a search which yields many results.
+  :bug:`5102`
 
 For packagers: